updated reducer to store data in state
